requirements.txt
Django>=4.2
djangorestframework>=3.14
drf-yasg>=1.21.7
psycopg2-binary>=2.9
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3